+++ src/components/forms/FieldWrapper.tsx
@@ -8,7 +8,7 @@ import {
   NecessityIndicator,
   ValidationState,
 } from '../../shared';
-import { Styles, tasty } from '../../tasty';
+import { Props, Styles, tasty } from '../../tasty';
 import { TooltipProvider } from '../overlays/Tooltip/TooltipProvider';
 import { InfoCircleOutlined } from '@ant-design/icons';
 import { wrapNodeIfPlain } from '../../utils/react';
@@ -25,13 +25,17 @@ const FieldElement = tasty({
       '': '1x',
       'has-sider': '@(column-gap, 1x)',
     },
-    placeItems: 'baseline stretch',
+    placeItems: 'start stretch',
 
     LabelArea: {
       display: 'block',
       width: {
         '': 'initial',
         'has-sider': '@label-width',
+      },
+      padding: {
+        '': 'initial',
+        'has-sider': '1.25x top',
       },
     },
 
@@ -73,8 +77,8 @@ export type CubeFieldWrapperProps = {
   /** Whether the input is disabled */
   isDisabled?: boolean;
   necessityIndicator?: NecessityIndicator;
-  labelProps?: any;
-  fieldProps?: any;
+  labelProps?: Props;
+  fieldProps?: Props;
   /** Custom message for the field. It will be placed below the label and the input */
   message?: string | ReactNode;
   /** Styles for the message */