+++ src/type-checks.tsx
@@ -1,0 +1,9 @@
+import { tasty } from './tasty';
+import { RadioButton } from './components/forms/RadioGroup/Radio';
+
+tasty(RadioButton, {
+  // value is required
+  inputStyles: {
+    fill: '#clear',
+  },
+});